IFS-2016-PF-SpendProfile-Action

+++ ifs-web-service/ifs-project-setup-mgt-service/src/main/java/org/innovateuk/ifs/project/status/populator/PopulatedCompetitionStatusViewModel.java
@@ -21,7 +21,7 @@ public class PopulatedCompetitionStatusViewModel {
     private final CompetitionProjectsStatusResource competitionProjectsStatus;
     private final CompetitionStatusViewModel viewModel;
 
-    public PopulatedCompetitionStatusViewModel(CompetitionProjectsStatusResource competitionProjectsStatus, UserResource user, long openQueryCount, long pendingSpendProfilesCount, boolean showTabs) {
+    public PopulatedCompetitionStatusViewModel(CompetitionProjectsStatusResource competitionProjectsStatus, UserResource user, long openQueryCount, int pendingSpendProfilesCount, boolean showTabs) {
         this.user = user;
         this.competitionProjectsStatus = competitionProjectsStatus;
         final boolean canExportBankDetails = user.hasRole(PROJECT_FINANCE);
